fix: add observer

+++ async-race/src/pages/garage/cars-list/cars-list.ts
@@ -33,6 +33,7 @@ export class CarsList implements Component {
   };
 
   private handleClick = async ({ target }: Event): Promise<void> => {
+    const { winners } = this.storeService.getState();
     const elem = target as HTMLElement;
 
     if (elem.tagName === 'UL') {
@@ -47,7 +48,9 @@ export class CarsList implements Component {
       UtilService.fillCarUpdate(car);
     } else if (targetId === ContentConstants.REMOVE) {
       await ApiService.deleteCar(carId);
-      await ApiService.deleteWinner(carId);
+      if (winners.find(({ id }) => id === carId) !== undefined) {
+        await ApiService.deleteWinner(carId);
+      }
       this.redrawPage();
     } else if (targetId === `${ContentConstants.START_BTN}-${carId}`) {
       UtilService.startDriving(carId);
